chore(suite-native): add styles provider
