+++ packages/suite-native/package.json
@@ -15,6 +15,8 @@
     },
     "dependencies": {
         "@trezor/transport-native": "*",
+        "@trezor/styles": "*",
+        "@trezor/theme": "*",
         "node-libs-browser": "^2.2.1",
         "react": "17.0.2",
         "react-native": "0.68.0",
