+++ packages/suite-native/tsconfig.json
@@ -2,7 +2,9 @@
     "extends": "../../tsconfig.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
-        { "path": "../transport-native" }
+        { "path": "../transport-native" },
+        { "path": "../styles" },
+        { "path": "../theme" }
     ],
     "include": [".", "**.json"]
 }
